Add files via upload
Re-implement in Python the results presented in Figure 6.4 of the Sutton & Barto book comparing SARSA 
and Q-learning in the cliff-walking task. Investigate the effect of choosing different values for the exploration 
parameter    for  both  methods.  Present  your  code  and  results.  In  your  discussion  clearly  describe  the  main 
difference between SARSA and Q-learning in relation to your findings.